requirements.dev.txt
fake-bpy-module-latest
requests
